requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
Flask-WTF>=1.0
Flask-Login>=0.5
Flask-Mail>=0.9
Flask-Bcrypt>=1.0  # Or use Werkzeug's security helpers
python-dotenv>=0.19
email-validator>=1.1 # Required by Flask-WTF for Email validation
Werkzeug>=2.0 # Usually comes with Flask, ensure recent version
WTForms-SQLAlchemy>=0.3